+++ src/pages/species-richness/species-richness.ts
@@ -20,16 +20,40 @@ dayjs.extend(utc)
 export default class SpeciesRichnessPage extends Vue {
   public streams: StreamModels.Stream[] = []
 
-  public chartData: ChartModels.BarChartItem[] = []
+  public chartData: ChartModels.GroupedBarChartItem[] = []
   mapDatasets: ChartModels.MapDataSet[] = []
 
-  onFilterChange (filters: SpeciesRichnessFilter[]): void {
-    for (const filter of filters) {
-      const start = filter.startDate.toISOString()
-      const end = filter.endDate.add(1, 'days').toISOString()
-      const data = SpeciesService.getMockupSpecies({ start, end, streams: filter.streams })
-      this.chartData = data
-    }
+  async onFilterChange (filters: SpeciesRichnessFilter[]): Promise<void> {
+    const groupedItems: { [key: string]: ChartModels.GroupedBarChartItem } = {}
+    const chartItems = await Promise.all(filters.map(f => {
+      const start = f.startDate.toISOString()
+      const end = f.endDate.add(1, 'days').toISOString()
+      return SpeciesService.getMockupSpecies({ start, end, streams: f.streams })
+    }))
+
+    const categories = new Set(chartItems.flatMap(i => i.map(c => c.category)))
+    categories.forEach(cat => {
+      for (const [idx, item] of chartItems.entries()) {
+        const filter = filters[idx]
+        const siteName = filter.streams.length > 0 ? filter.streams.map(s => s.name).join(',') : 'All sites'
+        const matchedData = item.find(d => d.category === cat)
+        const seriesItem: ChartModels.BarChartItem = {
+          category: siteName,
+          frequency: matchedData?.frequency ?? 0,
+          color: filter.color
+        }
+        if (groupedItems[cat] !== undefined) {
+          groupedItems[cat].series.unshift(seriesItem)
+        } else {
+          groupedItems[cat] = {
+            group: cat,
+            series: [seriesItem]
+          }
+        }
+      }
+    })
+
+    this.chartData = Object.values(groupedItems)
 
     // TODO 41 - Merge this with the above once Nutto's branch is merged
     this.mapDatasets = filters.map(({ startDate, endDate, streams, color }) => ({